+++ src/views/QueryEditor/QueryEditor.tsx
@@ -1,115 +1,37 @@
 import React, { useEffect, useState } from 'react';
 import { QueryEditorProps } from '@grafana/data';
 import { CHDataSource } from '../../datasource/datasource';
-import {CHDataSourceOptions, CHQuery, EditorMode, TimestampFormat} from '../../types/types';
+import { CHDataSourceOptions, CHQuery, EditorMode } from '../../types/types';
 import { QueryHeader } from './components/QueryHeader/QueryHeader';
 import { QueryTextEditor } from './components/QueryTextEditor/QueryTextEditor';
 import { QueryBuilder } from './components/QueryBuilder/QueryBuilder';
-import SqlQuery from '../../datasource/sql-query/sql_query';
-import { Alert } from "@grafana/ui";
-import {useSystemDatabases} from "../hooks/useSystemDatabases";
-import {useAutocompleteData} from "../hooks/useAutocompletionData";
-
-const defaultQuery = 'SELECT $timeSeries as t, count() FROM $table WHERE $timeFilter GROUP BY t ORDER BY t';
-const DEFAULT_FORMAT = 'time_series';
-const DEFAULT_DATE_TIME_TYPE = TimestampFormat.DateTime;
-const DEFAULT_ROUND = '0s';
-const DEFAULT_INTERVAL_FACTOR = 1;
-
-function useFormattedData(query: CHQuery, datasource: CHDataSource): [string, string | null] {
-  useSystemDatabases(datasource)
-  useAutocompleteData(datasource)
-  const [formattedData, setFormattedData] = useState(query.query);
-  const [error, setError] = useState<string | null>(null);
-
-  useEffect(() => {
-    try {
-      if (datasource.options && datasource.templateSrv) {
-        const queryModel = new SqlQuery(query, datasource.templateSrv, datasource.options);
-        // @ts-ignore
-        const adHocFilters = datasource.templateSrv.getAdhocFilters(datasource.name);
-        const replaced = queryModel.replace(datasource.options, adHocFilters);
-        setFormattedData(replaced);
-        setError(null);
-      }
-    } catch (e: any) {
-      setError(e?.message);
-    }
-  }, [query, datasource.name, datasource.options, datasource.templateSrv]);
-
-  return [formattedData, error];
-}
+import { Alert } from '@grafana/ui';
+import { useQueryState } from './hooks/useQueryState';
+import { useFormattedData } from './hooks/useFormattedData';
+import { initializeQueryDefaults } from './helpers/initializeQueryDefaults';
 
 export function QueryEditor(props: QueryEditorProps<CHDataSource, CHQuery, CHDataSourceOptions>) {
   const {
- datasource, query, onChange, onRunQuery 
-} = props;
+    datasource,
+    query,
+    onChange,
+    onRunQuery
+  } = props;
   const isAnnotationView = !props.app;
   const initializedQuery = initializeQueryDefaults(query, isAnnotationView, datasource, onChange);
   const [formattedData, error] = useFormattedData(initializedQuery, datasource);
-  const [datasourceName] = useState(datasource.name);
-  const [datasourceUid] = useState(datasource.uid);
-  const [refId] = useState(query.refId);
   const [editorMode, setEditorMode] = useState(initializedQuery.editorMode || EditorMode.Builder);
-
-  useEffect(() => {
-    const accessKey = `dataStorage_${datasourceName}_${datasourceUid}_${refId}`;
-    // On component mount
-    const storedData = localStorage.getItem(accessKey);
-    if (storedData) {
-      const { name, timestamp } = JSON.parse(storedData);
-      const currentTime = new Date().getTime();
-      const timeDifference = (currentTime - timestamp) / 1000; // Convert milliseconds to seconds
-
-      if (timeDifference < 5) {
-        if (name !== accessKey) {
-          const initialQuery = {
-            ...query,
-            format: DEFAULT_FORMAT,
-            extrapolate: true,
-            skip_comments: true,
-            add_metadata: true,
-            dateTimeType: DEFAULT_DATE_TIME_TYPE,
-            round: DEFAULT_ROUND,
-            intervalFactor: DEFAULT_INTERVAL_FACTOR,
-            interval: '',
-            query: defaultQuery,
-            formattedQuery: query.query,
-            editorMode: EditorMode.Builder,
-            database: undefined,
-            table: undefined,
-            dateColDataType: undefined,
-            dateTimeColDataType: undefined,
-          };
-
-          onChange(initialQuery);
-        }
-      }
-    }
-
-    // On component unmount
-    return () => {
-      const dataToStore = {
-        name: accessKey,
-        timestamp: new Date().getTime()
-      };
-      localStorage.setItem(accessKey, JSON.stringify(dataToStore));
-    };
-    // eslint-disable-next-line
-  }, []);
-
+  useQueryState(query, onChange, datasource);
   const onSqlChange = (sql: string) => onChange({ ...initializedQuery, query: sql });
-
   const onFieldChange = (value: any) => onChange({ ...query, ...value });
-
-  const onTriggerQuery = () => onRunQuery()
+  const onTriggerQuery = () => onRunQuery();
 
   // @ts-ignore
   const adHocFilters = datasource.templateSrv.getAdhocFilters(datasource.name);
 
   useEffect(() => {
     if (adHocFilters.length > 0) {
-      onChange({...query, adHocFilters: adHocFilters })
+      onChange({ ...query, adHocFilters: adHocFilters });
     }
   }, [adHocFilters.length]);
 
@@ -124,9 +46,14 @@ export function QueryEditor(props: QueryEditorProps<CHDataSource, CHQuery, CHDat
         onTriggerQuery={onTriggerQuery}
         onChange={onChange}
       />
-      {error ? <Alert title={error} elevated style={{marginTop: "5px", marginBottom: "5px"}}/> : null}
+      {error ? <Alert title={error} elevated style={{ marginTop: '5px', marginBottom: '5px' }} /> : null}
       {editorMode === EditorMode.Builder && (
-        <QueryBuilder query={initializedQuery} datasource={datasource} onChange={(items: CHQuery) => onChange({...items})} onRunQuery={onTriggerQuery} />
+        <QueryBuilder
+          query={initializedQuery}
+          datasource={datasource}
+          onChange={(items: CHQuery) => onChange({ ...items })}
+          onRunQuery={onTriggerQuery}
+        />
       )}
       {editorMode === EditorMode.SQL && (
         <>
@@ -145,52 +72,4 @@ export function QueryEditor(props: QueryEditorProps<CHDataSource, CHQuery, CHDat
       )}
     </>
   );
-}
-
-function initializeQueryDefaults(query: CHQuery, isAnnotationView: boolean, datasource: any, onChange: any): CHQuery {
-  const initializedQuery = {
-    ...query,
-    format: query.format || DEFAULT_FORMAT,
-    extrapolate: query.extrapolate ?? true,
-    skip_comments: query.skip_comments ?? true,
-    add_metadata: query.add_metadata ?? true,
-    dateTimeType: query.dateTimeType,
-    round: query.round || DEFAULT_ROUND,
-    intervalFactor: query.intervalFactor || DEFAULT_INTERVAL_FACTOR,
-    interval: query.interval || '',
-    adHocFilters: query.adHocFilters || [],
-    query: query.query || defaultQuery,
-    formattedQuery: query.formattedQuery || query.query,
-    editorMode: (query.database && query.table) ? EditorMode.SQL : EditorMode.Builder,
-  };
-
-  if (datasource.defaultValues && !query.initialized) {
-    if (datasource.defaultValues.defaultDateTimeType && !initializedQuery.dateTimeType) {
-      initializedQuery.dateTimeType = datasource.defaultValues.defaultDateTimeType;
-    }
-
-    if (datasource.defaultValues.dateTime.defaultDateTime && initializedQuery.dateTimeType === TimestampFormat.DateTime && !initializedQuery.dateTimeColDataType) {
-      initializedQuery.dateTimeColDataType = datasource.defaultValues.dateTime.defaultDateTime;
-    }
-
-    if (datasource.defaultValues.dateTime.defaultDateTime64 && initializedQuery.dateTimeType === TimestampFormat.DateTime64 && !initializedQuery.dateTimeColDataType) {
-      initializedQuery.dateTimeColDataType = datasource.defaultValues.dateTime.defaultDateTime64;
-    }
-
-    if (datasource.defaultValues.dateTime.defaultDateDate32 && !initializedQuery.dateColDataType) {
-      initializedQuery.dateColDataType = datasource.defaultValues.dateTime.defaultDateDate32;
-    }
-
-    if (datasource.defaultValues.dateTime.defaultUint32 && initializedQuery.dateTimeType === TimestampFormat.TimeStamp  && !initializedQuery.dateTimeColDataType) {
-      initializedQuery.dateTimeColDataType = datasource.defaultValues.dateTime.defaultUint32;
-    }
-
-    onChange({ ...query, ...initializedQuery, initialized: true });
-  }
-
-  if (isAnnotationView) {
-    initializedQuery.format = 'ANNOTATION'
-  }
-
-  return initializedQuery
 }
